# task4/gin-backend/docs/swagger.yaml
definitions:
  controllers.CreatePostRequest:
    description: 创建新的文章
    properties:
      content:
        minLength: 10
        type: string
      title:
        maxLength: 200
        minLength: 3
        type: string
    required:
    - content
    - title
    type: object
info:
  contact: {}
paths:
  /api/v1/comments/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定评论（仅评论作者可操作）
      parameters:
      - description: 评论ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无权操作
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 评论不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 删除评论
      tags:
      - comments
  /api/v1/posts:
    get:
      consumes:
      - application/json
      description: 获取所有文章的列表
      produces:
      - application/json
      responses:
        "200":
          description: 文章列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取文章列表
      tags:
      - posts
  /api/v1/posts/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定文章
      parameters:
      - description: 文章ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无权操作
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 文章不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 删除文章
      tags:
      - posts
    get:
      consumes:
      - application/json
      description: 根据文章ID获取文章详细信息
      parameters:
      - description: 文章ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 文章详情
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 文章不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取文章详情
      tags:
      - posts
    put:
      consumes:
      - application/json
      description: 更新指定文章的内容
      parameters:
      - description: 文章ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新后的文章信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.CreatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无权操作
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 文章不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 更新文章
      tags:
      - posts
  /api/v1/posts/{id}/comments:
    get:
      consumes:
      - application/json
      description: 获取指定文章的所有评论
      parameters:
      - description: 文章ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 评论列表
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 文章不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取文章评论列表
      tags:
      - comments
  /api/v1/user:
    get:
      consumes:
      - application/json
      description: 获取当前登录用户的详细信息
      produces:
      - application/json
      responses:
        "200":
          description: 用户信息
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 获取当前用户信息
      tags:
      - users
swagger: "2.0"